fix contributing instructions to avoid conflicts
The old instructions caused all commits to happen on the fork's master branch which is not always up to date with this repository. This caused conflicts after the first PR is merged.

The new instructions use GitHub's "Prose new file" feature witch 1) creates the fork automatically and 2) always create a new up-to-date branch on the fork.